refactor: share command handling
